[#8,9] 예약현황 조회 1. url 형식 restful 하게 변경 2. 기관조회, 예약현황조회 서비스 따로 분리
